requirements_.txt
asgiref==3.5.0
Django==3.2.13
django-cors-headers==3.11.0
django-debug-toolbar==3.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
factory-boy==3.2.1
Faker==13.15.0
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
